+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add _Aspirations_ as text fields to Accursed sheets' biography tabs
 - Enable manually sorting some item types in actor sheets
 - Make embedded items available in objects in `rollData` using their identifier as key, e.g. `actor.system.skills.technology`
+- Derive number of injury dice into accursed system model, allowing effects to change them
 
 ### Changed